video-processing-backend/requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aioredis==2.0.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==4.4.2
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.1
fastapi-cli==0.0.5
fastapi-limiter==0.1.6
ffmpeg-python==0.2.0
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
keras==3.5.0
kombu==5.4.0
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==4.25.4
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
redis==5.1.0b7
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
starlette==0.38.2
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tf_keras==2.17.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.23.0
wcwidth==0.2.13
websockets==13.0
Werkzeug==3.0.4
wrapt==1.16.0
